The package is now compatible iOS 10 / MacOS 10_12 / tvOS v10 / WatchOS v3
Even if the package is compatible, it doesn’t mean the composable architecture is available for these targets. However you can import the module and use it in your app components marked annotated with @available(iOS 13, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
The goal is for developers to be able to develop new features for iOS 13 while still maintaining a target of iOS 12 or below.